+++ packages/apollo-server-core/src/index.ts
@@ -20,3 +20,12 @@ export { convertNodeHttpToRequest } from './nodeHttpToRequest';
 // ApolloServer Base class
 export { ApolloServerBase } from './ApolloServer';
 export * from './types';
+
+//This currently provides the ability to have syntax highlighting as well as
+//consistency between client and server gql tags
+import { DocumentNode } from 'graphql';
+import gqlTag from 'graphql-tag';
+export const gql: (
+  template: TemplateStringsArray,
+  ...substitutions: any[]
+) => DocumentNode = gqlTag;