+++ packages/apollo-server-micro/src/microApollo.ts
@@ -1,11 +1,10 @@
 import {
   GraphQLOptions,
-  HttpQueryError,
   runHttpQuery,
   convertNodeHttpToRequest,
 } from 'apollo-server-core';
 import * as GraphiQL from 'apollo-server-module-graphiql';
-import { createError, json, RequestHandler } from 'micro';
+import { json, RequestHandler } from 'micro';
 import * as url from 'url';
 import { IncomingMessage, ServerResponse } from 'http';