+++ packages/apollo-server/src/exports.ts
@@ -1,8 +1,5 @@
 export * from 'graphql-tools';
 export * from 'graphql-subscriptions';
-// this makes it easy to get inline formatting and highlighting without
-// actually doing any work
-export const gql = String.raw;
 
 export {
   ApolloError,
@@ -11,4 +8,5 @@ export {
   ValidationError,
   AuthenticationError,
   ForbiddenError,
+  gql,
 } from 'apollo-server-core';